requirements.txt
Flask
Flask-Dance
Flask-SQLAlchemy
Flask-Login
SQLAlchemy-Utils
blinker
Flask-Migrate
Flask-Mail
